% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{docx_document}
\alias{docx_document}
\title{Convert to UNHCR branded Word document}
\usage{
docx_document(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{officedown::rdocx_document}}
}
\description{
Format for converting from R Markdown to an UNHCR branded Word document
}
