% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_simple.R
\name{docx_simple}
\alias{docx_simple}
\title{Convert to UNHCR branded Word document}
\usage{
docx_simple(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{officedown::rdocx_document}}
}
\value{
An R Markdown output format.
}
\description{
Format for converting from R Markdown to an UNHCR branded Word simple document
}
