% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_factsheet.R
\name{paged_factsheet}
\alias{paged_factsheet}
\title{Function for UNHCR paged Factsheet}
\usage{
paged_factsheet(other_css = NULL, number_sections = FALSE, ...)
}
\arguments{
\item{other_css}{Add extra css}

\item{number_sections}{Number section headings}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown report
}
\description{
Function for UNHCR paged Factsheet
}
