% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_presentation.R
\name{ppt_presentation}
\alias{ppt_presentation}
\title{Convert to UNHCR branded Powerpoint presentation}
\usage{
ppt_presentation(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{rmarkdown::powerpoint_presentation}}
}
\description{
Format for converting from R Markdown to an UNHCR branded Powerpoint presentation
}
