% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_presentation.R
\name{pptx_presentation}
\alias{pptx_presentation}
\title{Convert to UNHCR branded Powerpoint presentation}
\usage{
pptx_presentation(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{rmarkdown::powerpoint_presentation}}
}
\description{
Format for converting from R Markdown to an UNHCR branded Powerpoint presentation
}
