% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_presentation.R
\name{xaringan_presentation}
\alias{xaringan_presentation}
\title{Convert to UNHCR branded Xaringan presentation}
\usage{
xaringan_presentation(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{xaringan::moon_reader}}
}
\description{
Format for converting from R Markdown to an UNHCR branded Xaringan presentation
}
