% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_page.R
\name{html_page}
\alias{html_page}
\title{Convert to UNHCR branded HTML website}
\usage{
html_page(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{rmarkdown::html_document}}
}
\description{
Format for converting from R Markdown to an UNHCR branded HTML website
}
