% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_slides.R
\name{html_slides}
\alias{html_slides}
\title{Convert to UNHCR branded HTML presentation}
\usage{
html_slides(other_css = NULL, ...)
}
\arguments{
\item{other_css}{Add extra css}

\item{...}{extra parameters to pass to \code{xaringan::moon_reader}}
}
\value{
An R Markdown output format.
}
\description{
Format for converting from R Markdown to an UNHCR branded HTML presentation
}
