% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_slides.R
\name{html_slides}
\alias{html_slides}
\title{Convert to UNHCR branded HTML presentation}
\usage{
html_slides(
  other_css = NULL,
  nature = list(countIncrementalSlides = FALSE, highlightLines = TRUE, highlightStyle =
    "github", ratio = "16:9"),
  ...
)
}
\arguments{
\item{other_css}{Add extra css}

\item{nature}{See ?xaringan::moon_reader}

\item{...}{extra parameters to pass to \code{xaringan::moon_reader}}
}
\description{
Format for converting from R Markdown to an UNHCR branded HTML presentation
}
