% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slides.R
\name{pptx_slides}
\alias{pptx_slides}
\title{Convert to UNHCR branded Powerpoint presentation}
\usage{
pptx_slides(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{officedown::rpptx_document}}
}
\value{
An R Markdown output format.
}
\description{
Format for converting from R Markdown to an UNHCR branded Powerpoint presentation
}
